+++ app/scripts/controllers/statistics_controller.ts
@@ -119,7 +119,7 @@ angular.module("korpApp").directive("statsResultCtrl", () => ({
                     )
                     .catch((error) => {
                         // AbortError is expected if a new search is made before the previous one is finished
-                        if ((error.name = "AbortError")) return
+                        if (error.name == "AbortError") return
                         console.error(error)
                         // TODO Show error
                         $timeout(() => {